restructured working directory with folders
