+++ TMdllB/Movie/Movie.cs
@@ -1,4 +1,4 @@
-﻿namespace TMdllB
+﻿namespace TMdllB.Movie
 {
     public class Movie
     {
@@ -46,7 +46,9 @@
 
         public string RelativeBackdropPath { get; }
 
-        public string BackdropPath { get
+        public string BackdropPath
+        {
+            get
             {
                 return $"{TMDbClient.IMAGE_BASE_URL}{TMDbClient.IMAGE_ORIGINAL_SIZE_ENDPOINT}{RelativeBackdropPath}";
             }
@@ -62,7 +64,9 @@
 
         public string RelativePosterPath { get; }
 
-        public string PosterPath { get
+        public string PosterPath
+        {
+            get
             {
                 return $"{TMDbClient.IMAGE_BASE_URL}{TMDbClient.IMAGE_ORIGINAL_SIZE_ENDPOINT}{RelativePosterPath}";
             }
